Add files via upload
In this Employee Attendance Monitoring project, attempts were made to register the presence of employees. The main purpose of this java application is to maintain information about the presence that all employees can use in their departments. This web application retrieves the details of the departments assigned to the relevant administrative employees and employees enrolled in the courses from the server with the internet connection and stores them in the database.